dbt_artifacts_based/chatdbt-lmstudio/backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-dotenv==1.1.0
pydantic==2.11.7
llama-index==0.12.42
llama-index-core==0.12.42
llama-index-llms-openai==0.4.7
llama-index-embeddings-openai==0.3.1
llama-index-readers-file==0.4.9
openai==1.88.0
aiofiles==24.1.0
aiosqlite==0.21.0
anyio==4.9.0
httpx==0.28.1
tiktoken==0.9.0
numpy==2.2.6
pandas==2.2.3
networkx==3.4.2
sentence-transformers==3.3.1
